# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TravelBuddi API",
        "description": "A product of Appetiser",
        "contact": {
            "name": "Iazell",
            "email": "iazell.raper@appetiser.com.au"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{ip or domain}/api/v1",
            "description": "Base Endpoint"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retrieve user's information",
                "operationId": "show",
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "User Not Found / Token Not Found"
                    },
                    "422": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user's information",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string",
                                        "example": "Philippines"
                                    },
                                    "mobile_number": {
                                        "description": "Mobile Number",
                                        "type": "string",
                                        "example": "+63 9277468888"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string",
                                        "example": "216 Cabreros Street Cebu City"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "is_allowed": {
                                        "description": "If the user is allowed to post prices",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "User Not Found / Token Not Found"
                    },
                    "422": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/user/photo": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Upload profile photo",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiUsersController::uploadProfilePhoto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "description": "Image File",
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "User Not Found / Token Not Found"
                    },
                    "422": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}